split statusbar into multiple files
