+++ frontend/javascripts/viewer/view/layouting/flex_layout_wrapper.tsx
@@ -40,7 +40,7 @@ import ConnectomeView from "viewer/view/right_border_tabs/connectome_tab/connect
 import DatasetInfoTabView from "viewer/view/right_border_tabs/dataset_info_tab_view";
 import SegmentsView from "viewer/view/right_border_tabs/segments_tab/segments_tab_view";
 import SkeletonTabView from "viewer/view/right_border_tabs/skeleton_tab/skeleton_tab_view";
-import Statusbar from "viewer/view/statusbar";
+import Statusbar from "viewer/view/statusbar/statusbar";
 import TDViewControls from "viewer/view/td_view_controls";
 import BorderToggleButton from "../components/border_toggle_button";
 import type {
